+++ trunk/simulations-java/common/piccolo-phet/src/edu/colorado/phet/common/piccolophet/nodes/layout/HBox.java
@@ -31,7 +31,7 @@ public class HBox extends Box {
     }
 
     //Creates a VBox which lays out nodes horizontally.  This constructor invocation is meant to be read with 'code folding' on and a good healthy right margin (like 200)
-    public HBox( int spacing,
+    public HBox( double spacing,
                  PNode... children//List of children to be added on initialization
     ) {
         super( spacing,
